+++ src/components/tooltip/tooltip_constants.js
@@ -1,0 +1,6 @@
+export const SIZE = {
+  SMALL: 's',
+  MEDIUM: 'm',
+  LARGE: 'l',
+  AUTO: 'auto',
+};
